- separate test for publisher from parser
