+++ src/test/resources/hudson/plugins/checkstyle/testXxe-xxe.xml
